==> PlayStore/2.1- Data Preparation - Train.csv.ipynb <==
# 2.1- Data Preparation - Train.csv

import pandas as pd
from matplotlib import pyplot as plt
import seaborn as sns
import numpy as np

import warnings
warnings.filterwarnings('ignore')

df = pd.read_csv("datasets/TrainDataAnalysis.csv")

df.head()

## Feature Engineering content

* Delete `App` column, because is different in each row
* Change `Size` column values to millions or thousands
* Delete `Install` signature of + 
* Delete `$` symbol to Price
* Dummy encode `Categories, Content Rating, Genres`, because they don't have any range as ordinal data
* Last updated change to one column with year, one column with month, one column with day
* Delete `Current Ver`, because don't give any value
* In `Android Ver` edit number deleting the word "and up", and deleting the outliers with forma "-"
* Boolean values for `Type`
* Give an ordinal number to `review_group` as range of importance
* Delete `App, Size, Installs, Price, Last Updated, LastUpdatedNew, Current Ver, Android Ver, review_group` because is already encoded
* Impute Null values
* Normalization, Standarization???

### 1. Personalized Changes

#### 1.1 Changing Size format
From values with M and k, to floats

def change_M_k(df, new_column_name, column_name):
    df[new_column_name] = df[column_name].apply(lambda x: 
                         float(x.split("M")[0]) / 100 if x[-1] == "M" else 
                         float(x.split("k")[0]) / 100 if x[-1] == "k" else
                         np.nan
                    )
    return df

df = change_M_k(df, "SizeNew", "Size")

df.head()

#### 1.2 Changing installs "+" format for integers

def delimiter_format(df, new_column_name, column_name):
    df[new_column_name] = df[column_name].apply(lambda x: x.split("+")[0] if x[-1] == "+" else np.nan)
    df[new_column_name] = df[new_column_name].apply(lambda x: int(x.replace(',', '')))
    
    return df

df = delimiter_format(df, 'InstallsNew', 'Installs')

df.head()

#### 1.3 Changing symbol "$" in Pricing

def delete_price_symbol(df, new_column_name, column_name):
    df[new_column_name] = df[column_name].apply(lambda x: float(x.split("$")[1]) if x[0] == "$" else 0)
    return df

df = delete_price_symbol(df, "PriceNew", "Price")

df.head()

#### 1.4 Encoding Dates, separating each year, month and weekday

def change_date_format(df, new_column_name, column_name):
    df[new_column_name] = pd.to_datetime(df[column_name])
    df["{}Year".format(new_column_name)] = df[new_column_name].dt.year
    df["{}Month".format(new_column_name)] = df[new_column_name].dt.month
    df["{}Day".format(new_column_name)] = df[new_column_name].dt.day
    return df

df = change_date_format(df, "LastUpdatedNew", "Last Updated")

df.head()

#### 1.5 Converting semversion to a float

def semversion_to_number(df, new_column_name, column_name):
    ## Find rare characters
    df[new_column_name] = df[column_name].apply(lambda x: 
                            x.split("W")[0] if x[-8] == "W" else
                            x.split("-")[0] if x[-7] == "-" else
                            x.split(" and up")[0] if x[-1] == "p" else
                            str(0)) ## change later this cero for NaN
    
    ## Find records with 3 poinbt version, like 4.0.2 and replace for float with first version
    df[new_column_name] = df[new_column_name].apply(lambda x: 
                            float(x.split(".")[0]) if x.count('.') == 2 else
                            float(x))
    
    ## Change zeros for NaNs
    df[new_column_name] = df[new_column_name].apply(lambda x: 
                            np.nan if x == 0 else
                            x)
    
    return df

df = semversion_to_number(df, "AndroidVerNew", "Android Ver")

df.head()

#### 1.6 Encoding review groups to oridinal numbers

def from_review_group_to_oridinal(df, new_column_name, column_name):
    df[new_column_name] = df[column_name].apply(
        lambda x: 0 if x == '[0, 23)' else 
        1 if x == '[23, 889)' else
        2 if x == '[889, 26432)' else
        3
    )
    return df

df = from_review_group_to_oridinal(df, "ReviewGroupNew", "review_group")

df.head()

### 2- Generic Changes

#### 2.1 Null values imputation

**Note:** Please note here that this data is for a competition, so I don't need to worry about data leakage

def null_exploration(df):
    ## to explore nulls
    print(df.isnull().sum())
    print(df.isnull().mean())

null_exploration(df)

def null_imputation(column_name):
    df[column_name].fillna(df[column_name].mean(), inplace=True)
    return df

df = null_imputation("SizeNew")

df = null_imputation("AndroidVerNew")

#### 2.2 OneStepFeatEng

#### One step feature engineering

class OneStepFeatEng():
    def __init__(self, df):
        self.df = df
        
    # Drop columns
    def drop_column(self, column_names):
        self.df.drop(column_names, axis=1, inplace=True)
        return self.df.head()
    
    ## Separated cols
    def separated_cols(self, target_var):
        target_var = target_var
        features = [x for x in list(self.df.columns) if x != target_var]
        
        cat_cols = [column_name for column_name in self.df.columns if self.df[column_name].dtypes=='O']
        num_cols = [column_name for column_name in self.df.columns if self.df[column_name].dtypes!='O']
        
        return features, cat_cols, num_cols
    
    # Boolean encoding
    def boolean_encoding(self, column_name, true_val):
        self.df[column_name] = self.df[column_name].apply(lambda x: 1 if x == true_val else 0)
        return self.df.head()
    
    # Final columns
    def final_columns(self):
        return self.df.columns
    
    # Final dataframe
    def final_dataframe(self):
        return self.df

## Instance
X = OneStepFeatEng(df)

X.drop_column(['App', 'Size', 'Installs', 'Price', 'Last Updated', 'LastUpdatedNew', 'Current Ver', 
               'Android Ver', 'review_group'])

X.boolean_encoding("Type", "Paid")

X.final_dataframe().Type.value_counts()

features, cat_cols, num_cols = X.separated_cols('Rating')

X.final_dataframe().head()

features

cat_cols

num_cols

X.final_dataframe().to_csv("datasets/TrainFeatureEngined.csv", index=False)

X.final_dataframe()
