==> PlayStore/2.3- Data Preparation - All.ipynb <==
# 2.3- Data Preparation - All

Get Dummies for Both (Train + Test)

import pandas as pd
import numpy as np

Train =  pd.read_csv("datasets/TrainFeatureEngined.csv")
Test =  pd.read_csv("datasets/TestFeatureEngined.csv")

Train.head()

Train.shape

Test.head()

Test.shape

# I have to do this because train and test datasets finalice with a different number 
# of columns after doing dummy encopding with datasets separated
# im following this technicque: https://medium.com/@vaibhavshukla182/how-to-solve-mismatch-in-train-and-test-set-after-categorical-encoding-8320ed03552f

class GetDummies():
    
    def __init__(self, train, test):
        self.train = train
        self.test = test
    
    # Dummy encoding    
    def dummy_encoding(self, column_name):
        self.train['train'] = 1
        self.test['train'] = 0
        
        combined = pd.concat([self.train, self.test])
        column_encoded = pd.get_dummies(combined[column_name])
        column_encoded.columns = ['{}_{}'.format(column_name, x) for x in column_encoded.columns]
        
        combined = pd.concat([combined, column_encoded], axis=1)
        
        train = combined[combined['train'] == 1]
        test = combined[combined['train'] == 0]
        
        train.drop(['train'], axis=1, inplace=True)
        test.drop(['train'], axis=1, inplace=True)
        
        return train, test

get_dummies_instance = GetDummies(Train, Test)

train_encoded, test_encoded = get_dummies_instance.dummy_encoding('Category')

train_encoded.head()

test_encoded.head()

train_encoded, test_encoded = get_dummies_instance.dummy_encoding('Content Rating')

train_encoded, test_encoded = get_dummies_instance.dummy_encoding('Genres')

train_encoded.head()

test_encoded.head()

### Drop Original Columns

train_encoded.drop(['Category', 'Content Rating', 'Genres'], axis=1, inplace=True)

train_encoded.head()

train_encoded.shape

test_encoded.drop(['Category', 'Content Rating', 'Genres', 'Rating'], axis=1, inplace=True)

test_encoded.head()

test_encoded.shape

train_encoded.to_csv("datasets/TrainEncoded.csv", index=False)
test_encoded.to_csv("datasets/TestEncoded.csv", index=False)
