==> PlayStore/6 - Modeling with Resample.ipynb <==
import pandas as pd

## modelos
from sklearn.neighbors import KNeighborsClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.linear_model import RidgeClassifier
from sklearn.linear_model import SGDClassifier
from sklearn.linear_model import PassiveAggressiveClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.tree import ExtraTreeClassifier
from sklearn.svm import LinearSVC
from sklearn.svm import SVC
from sklearn.naive_bayes import GaussianNB
from sklearn.ensemble import AdaBoostClassifier
from sklearn.ensemble import BaggingClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import ExtraTreesClassifier
from sklearn.gaussian_process import GaussianProcessClassifier
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis
from lightgbm import LGBMClassifier
from catboost import CatBoostClassifier
from xgboost import XGBClassifier

X_oversample = pd.read_csv('datasets/X_oversample.csv')
y_oversample = pd.read_csv('datasets/y_oversample.csv')
Test = pd.read_csv('datasets/TestEncoded.csv')

def get_datasets(train, test, target_var):
    y_train = Train[[target_var]]
    X_train = Train.drop(target_var, axis = 1)
    X_test = Test
    print(len(X_train), len(X_train.columns), X_train.columns)
    return X_train, y_train, X_test

X_train, y_train, X_test = X_oversample, y_oversample, Test

from sklearn.neighbors import KNeighborsClassifier

model = KNeighborsClassifier()

def get_models():
    models = dict() #list() # append
    models.update({'KNeighborsClassifier': KNeighborsClassifier()})
    models.update({'LogisticRegression': LogisticRegression()})
    models.update({'RidgeClassifier': RidgeClassifier()})
    models.update({'SGDClassifier': SGDClassifier()})
    models.update({'PassiveAggressiveClassifier': PassiveAggressiveClassifier()})
    models.update({'DecisionTreeClassifier': DecisionTreeClassifier()})
    models.update({'ExtraTreeClassifier': ExtraTreeClassifier()})
    models.update({'LinearSVC': LinearSVC()})
    models.update({'SVC': SVC()})
    models.update({'GaussianNB': GaussianNB()})
    models.update({'AdaBoostClassifier': AdaBoostClassifier()})
    models.update({'BaggingClassifier': BaggingClassifier()})
    models.update({'RandomForestClassifier': RandomForestClassifier()})
    models.update({'ExtraTreesClassifier': ExtraTreesClassifier()})
    models.update({'GaussianProcessClassifier': GaussianProcessClassifier()})
    models.update({'GradientBoostingClassifier': GradientBoostingClassifier()})
    models.update({'LinearDiscriminantAnalysis': LinearDiscriminantAnalysis()})
    models.update({'QuadraticDiscriminantAnalysis': QuadraticDiscriminantAnalysis()})
    models.update({'LGBMClassifier': LGBMClassifier()})
    models.update({'CatBoostClassifier': CatBoostClassifier()})
    models.update({'XGBClassifier': XGBClassifier()})
    return models
get_models()

def results_to_csv(file_name, y_pred):
    model = pd.DataFrame({'id':list(range(1,len(y_pred)+1)),
                        'rating': y_pred})
    model.to_csv(f'results/{file_name}.csv', index = False)
    return model
def modeling(X_train, y_train, X_test, num_iter):
    #X_train, y_train, X_test = get_datasets(train, test, target_var)
    models = get_models()
    for key,value in models.items():
        clf = value.fit(X_train, y_train)
        y_pred = clf.predict(X_test).reshape(-1,)
        results_to_csv(f'{key} - {num_iter}', y_pred)
    return y_pred

modeling(X_train, y_train, X_test, 2)

X_over_under_sample = pd.read_csv('datasets/X_over_under_sample.csv')
y_over_under_sample = pd.read_csv('datasets/y_over_under_sample.csv')
X_test = pd.read_csv('datasets/TestEncoded.csv')
X_train, y_train = X_over_under_sample, y_over_under_sample
len(X_test)

modeling(X_train, y_train, X_test, 3)

lin = pd.read_csv('results/LinearSVC - 1.csv')

lin.rating = lin.rating.apply(lambda x: 0 if x == 1 else 1)

lin.to_csv('results/LinearSVC - 1 inv.csv', index = False)

gauss

    models = dict() #list() # append
    models.update({'KNeighborsClassifier': KNeighborsClassifier()})
    models.update({'LogisticRegression': LogisticRegression()})
    models.update({'RidgeClassifier': RidgeClassifier()})
    models.update({'SGDClassifier': SGDClassifier()})
    models.update({'PassiveAggressiveClassifier': PassiveAggressiveClassifier()})
    models.update({'DecisionTreeClassifier': DecisionTreeClassifier()})
    models.update({'ExtraTreeClassifier': ExtraTreeClassifier()})
    models.update({'LinearSVC': LinearSVC()})
    models.update({'SVC': SVC()})
    models.update({'GaussianNB': GaussianNB()})
    models.update({'AdaBoostClassifier': AdaBoostClassifier()})
    models.update({'BaggingClassifier': BaggingClassifier()})
    models.update({'RandomForestClassifier': RandomForestClassifier()})
    models.update({'ExtraTreesClassifier': ExtraTreesClassifier()})
    models.update({'GaussianProcessClassifier': GaussianProcessClassifier()})
    models.update({'GradientBoostingClassifier': GradientBoostingClassifier()})
    models.update({'LinearDiscriminantAnalysis': LinearDiscriminantAnalysis()})
    models.update({'QuadraticDiscriminantAnalysis': QuadraticDiscriminantAnalysis()})
    models.update({'LGBMClassifier': LGBMClassifier()})
    models.update({'CatBoostClassifier': CatBoostClassifier()})
    models.update({'XGBClassifier': XGBClassifier()})

import glob

filenames = glob.glob('results/*')
frr= dict()
for i,j in filenames,models:
    dict
