==> PlayStore/8- Hyperparameter Tunning.ipynb <==
# 8- Hyperparameter Tunning

### def name_f(parm1, parm2, param3)
    pass
    
**1- Model parameter**

    * No es una entrada del usuario
    * Es un parametro resultado del modelo en sí (por ej. la pendiente o el intercepto)

**2- Model Hyperparameter**

    * Es una entrada del usuario
    * Learning rate / K nearest neighbors
    
### **2 opciones para hacer Hyperparameter Tunning:**

    1- Random Search (sklearn)
        ## KNN == Vecinos más cercanos (5,1,80,15)
    2- Grid Search
        ##KNN == Vecinos más cercanos range(1, 50, 3)

from sklearn.neighbors import KNeighborsClassifier
from sklearn.datasets import make_classification

from sklearn.neighbors import KNeighborsClassifier
from sklearn.datasets import make_classification

import pandas as pd
# def get_datasets():
#     X, y = make_classification(n_samples=1000, n_features=20, n_informative=15, 
#                               n_redundant=5, random_state=1)
#     return X, y
# X,y = get_datasets()

from sklearn.model_selection import KFold
from sklearn.model_selection import GridSearchCV, RandomizedSearchCV

##definir los hyperparametros del modelo
# model = KNeighborsClassifier()
# n_neighbors = range(1, 20)
# weights = ['uniform', 'distance']
# distance = ['euclidean', 'manhattan', 'minkowski']

##definir GridSearch

# grid = dict(n_neighbors = n_neighbors, weights = weights, metric = distance)
# cv = KFold(n_splits=5)
# grid_search = GridSearchCV(estimator = model, param_grid = grid, n_jobs=-1, cv=cv, scoring= 'f1')
# grid_result = grid_search.fit(X,y)

# grid_search

# grid_result.best_score_

# grid_result.best_params_

X_oversample = pd.read_csv('datasets/X_oversample.csv')
y_oversample = pd.read_csv('datasets/y_oversample.csv')
X_test = pd.read_csv('datasets/TestEncoded.csv')

from xgboost import XGBClassifier

def grid_search(params, random = False):
    
    ##XGBoost standard
    xgb = XGBClassifier(booster = 'gbtree', objective= 'binary:logistic', random_state = 2)
    
    ##KFold
    kfold = KFold(n_splits=5)
    
    ## if statement ocn random o sin random
    if random: 
        grid = RandomizedSearchCV(xgb, params, cv = kfold, n_iter=5, n_jobs=-1)
    else:
        grid = RandomizedSearchCV(xgb, params, cv = kfold , n_jobs=-1)
        
    ## entrenar el modelo en el grid
    grid.fit(X_oversample, y_oversample)
    
    ## obtener e imprimir los mejores parametros
    best_params= grid.best_params_
    print(best_params)
    
    ##obtener e imprimir mejor score
    best_score = grid.best_score_
    
    print('best score: ', best_score)

def grid_search(params, random=False):
    ## XGBoost standard
    xgb = XGBClassifier(booster='gbtree', objective='binary:logistic', random_state=2)
    ## Kfold
    kfold = KFold(n_splits=5)
    ## if statement con Random o sin Random
    if random:
        grid = RandomizedSearchCV(xgb, params, cv=kfold, n_iter=5, n_jobs=-1)
    else:
        grid = GridSearchCV(xgb, params, cv=kfold, n_jobs=-1)
    ## entrenar el modelo en el grid
    grid.fit(X_oversample, y_oversample)
    ## obtener e imprimir los mejores parametros
    best_params = grid.best_params_
    print(best_params)
    ## obtener e imprimir los mejores scores
    best_score = grid.best_score_
    print("best score", best_score)

grid_search(params = {'n_stimators': [2,25,50,75,100] })

grid_search(params = {'max_depth': [1,2,3,4,5,6,7,8], 'n_estimators':[100]})

grid_search(params = {'learning_rate':[0.01,0.05, 0.1, 0.2, 0.3, 0.4, 0.5], 'max_depth': [8], 'n_estimators':[100]})

grid_search(params = {'min_child_weight': [1,2,3,4,5],'learning_rate':[0.5], 'max_depth': [8], 'n_estimators':[100]})

model= XGBClassifier(booster= 'gbtree',objective='binary:logistic', random_state= 2 ,min_child_weight = 1,learning_rate=0.5 , max_depth= 8, n_estimators= 100)

model.fit(X_oversample,y_oversample)

y_pred = model.predict(X_test)

def results_to_csv(file_name, y_pred):
    model = pd.DataFrame({'id':list(range(1,len(y_pred)+1)),
                        'rating': y_pred})
    model.to_csv(f'results/{file_name}.csv', index = False)
    return model

results_to_csv("XGBoost-HyperparameterTunning-2", y_pred)
